requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
asttokens==3.0.0
certifi==2024.12.14
charset-normalizer==3.4.1
decorator==5.1.1
distro==1.9.0
Django==5.1.5
django-cors-headers==4.6.0
django-filter==24.3
djangorestframework==3.15.2
executing==2.1.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
ipython==8.25.0
jedi==0.19.2
jiter==0.8.2
matplotlib-inline==0.1.7
openai==1.59.8
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
prompt_toolkit==3.0.48
psycopg2==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.10.5
pydantic_core==2.27.2
Pygments==2.19.1
PyMuPDF==1.25.3
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
sqlparse==0.5.3
stack-data==0.6.3
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.3.0
wcwidth==0.2.13
whitenoise==6.8.2
